refactor(index): change index to be an constructor
hoodie/client is now a default setup.

BREAKING CHANGE: Export is now a constructor.
